renamed utils temporarily, updated tests
