+++ .github/workflows/python-test.yaml
@@ -34,6 +34,29 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    - name: Run tests
+    - name: Run tests and generate coverage report
       run: |
-        pytest --cov --cov-report=html
+        pytest --cov=src/crimson/templator --cov-report=html
+    - name: Upload coverage report
+      uses: actions/upload-artifact@v3
+      with:
+        name: coverage-report
+        path: htmlcov
+
+  notify-coverage:
+    runs-on: ubuntu-latest
+    needs: test-and-coverage
+    if: success() || failure()
+    steps:
+      - name: Download coverage report
+        uses: actions/download-artifact@v3
+        with:
+          name: coverage-report
+          path: htmlcov
+
+      - name: Post coverage report link
+        run: |
+          echo "Coverage report: ${GITHUB_SERVER_URL}/${GITHUB_REPOSITORY}/actions/runs/${GITHUB_RUN_ID}" > coverage_report.txt
+          gh pr comment ${{ github.event.pull_request.number }} --body-file coverage_report.txt
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
