+++ test/test_format_insert_loop.py
