+++ test/test_remove_lines.py
